auth

+++ modules/posts/posts.controller.js
@@ -1,7 +1,21 @@
 const Post = require("./posts.models");
 
+const verifyAuthor = async (req, user) => {
+  let post = await Post.findById(req.params.postId);
+  if (post._id.toString() !== user.req.id) {
+    return res
+      .status(406)
+      .json({ err: "You are not permitted to perform this operation" });
+  }
+};
+
 const getPost = async (req, res) => {
   const posts = await Post.find({});
+  res.status(200).json({ posts });
+};
+
+const getAllPostsByUser = async (req, res) => {
+  const posts = await Post.find({ author: req.user.id });
   res.status(200).json({ posts });
 };
 
@@ -10,7 +24,7 @@ const createPost = async (req, res) => {
   const post = await Post.create({
     title,
     body,
-    published,
+    author: req.user.id,
   });
   res.status(201).json({ post });
 };
@@ -23,7 +37,15 @@ const getSinglePost = async (req, res) => {
 
 const updatePost = async (req, res) => {
   const { postId } = req.params;
-  const post = await Post.findByIdAndUpdate(
+  //checks
+  await verifyAuthor();
+  let post = await Post.findById(req.params.postId);
+  if (post._id.toString() !== user.req.id) {
+    return res
+      .status(406)
+      .json({ err: "You are not permitted to perform this operation" });
+  }
+  post = await Post.findByIdAndUpdate(
     postId,
     { ...req.body },
     {
@@ -35,6 +57,13 @@ const updatePost = async (req, res) => {
 
 const deletePosts = async (req, res) => {
   const { postId } = req.params;
+  await verifyAuthor();
+  let post = await Post.findById(req.params.postId);
+  if (post._id.toString() !== user.req.id) {
+    return res
+      .status(406)
+      .json({ err: "You are not permitted to perform this operation" });
+  }
   await Post.findByIdAndDelete(postId);
   res.status(201).json({ msg: "Post deleted" });
 };
@@ -45,4 +74,5 @@ module.exports = {
   getSinglePost,
   getPost,
   createPost,
+  getAllPostsByUser,
 };
